pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "queuereader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.11.1",
    "pika>=1.3.2",
    "python-dotenv>=1.0.0",
    "setuptools>=80.9.0",
]

[project.scripts]
queuereader = "queuereader.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["main.py", "*.py"]

[tool.setuptools]
py-modules = []

